Better management of enemy mode
I'm trying something whereby there's a current mode and a requested mode (from the world).
Also, there's a "step out" mode to step outside of the ghost house but, frankly, this is getting a bit too unwieldy.
I'm about to experiment with an A* star algorithm that will take the ghost house's "wall" automatically into account when trying to come out of it and also not step into it when chasing heroes.