Get rid of whitespace in level information
I finally decided to get rid of the extra whitespace at the top and bottom of the maze asset.
It was originally there because the maze was generated from an auto-generator that would, I believe, produce larger mazes.
I'll stick to the current size for now and assume it will remain the same for the long term.